+++ packages/article-image/__tests__/web/jest.config.js
@@ -1,10 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("article-image", "web"), {
-  collectCoverageFrom: [
-    "**/packages/article-image/*.js",
-    "!**/packages/article-image/article-image.js",
-    "!**/packages/article-image/*stories.js",
-    "!**/packages/article-image/inset-caption.js"
-  ]
-});
+module.exports = jestConfigurator("article-image", "web");