+++ packages/image/__tests__/android/jest.config.js
@@ -1,12 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("image", "android"), {
-  collectCoverageFrom: [
-    "**/packages/image/**/*.js",
-    "!**/packages/image/**/image.js",
-    "!**/packages/image/**/*.web.js",
-    "!**/packages/image/**/*stories*",
-    "!**/packages/image/__tests__/**",
-    "!**/packages/image/coverage/**"
-  ]
-});
+module.exports = jestConfigurator("image", "android");