+++ packages/card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.10"></a>
+## [0.16.10](https://github.com/newsuk/times-components/compare/@times-components/card@0.16.9...@times-components/card@0.16.10) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
 <a name="0.16.9"></a>
 ## [0.16.9](https://github.com/newsuk/times-components/compare/@times-components/card@0.16.8...@times-components/card@0.16.9) (2018-01-05)