+++ packages/gradient/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.7"></a>
+## [0.3.7](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.3.6...@times-components/gradient@0.3.7) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/gradient
+
 <a name="0.3.6"></a>
 ## [0.3.6](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.3.5...@times-components/gradient@0.3.6) (2018-01-04)